extracted navbar to a Nav component
