+++ src/App.tsx
@@ -1,6 +1,5 @@
 import React, { useState, useEffect } from 'react';
-import { Switch, Route, useHistory } from 'react-router-dom';
-import { HStack, Button } from '@chakra-ui/react';
+import { Switch, Route } from 'react-router-dom';
 
 import { Login } from './features/auth/Login';
 import { Signup } from './features/auth/Signup';
@@ -10,14 +9,13 @@ import { useAuth } from './hooks/useAuth';
 import { useVerifyTokenQuery } from './app/services/api';
 import { useAppDispatch } from './hooks/store';
 import { setCredentials, resetCredentials } from './features/auth/authSlice';
-import { resetTodos } from './features/todo/todoSlice';
 
+import Nav from './features/nav/Nav';
 import AddTodo from './features/todo/AddTodo';
 import TodoList from './features/todo/TodoList';
 
 const App = (): JSX.Element => {
   const [allSetup, setAllSetup] = useState(false);
-  const { push } = useHistory();
   const { user } = useAuth();
   const { data, isLoading } = useVerifyTokenQuery();
   const dispatch = useAppDispatch();
@@ -41,35 +39,9 @@ const App = (): JSX.Element => {
     }
   }, [user]);
 
-  const logoutHandler = () => {
-    dispatch(resetTodos());
-    dispatch(resetCredentials());
-    localStorage.removeItem('token');
-    push('/login');
-  };
-
   return (
     <div className="App">
-      <header className="App-header">
-        <h1>Welcome to Collaborative Tasks!</h1>
-        {user ? (
-          <>
-            <p>Hello {user ? user.username : null}</p>
-            <Button h="1.75rem" size="sm" onClick={logoutHandler}>
-              Logout
-            </Button>
-          </>
-        ) : (
-          <HStack>
-            <Button h="1.75rem" size="sm" onClick={() => push('/login')}>
-              Login
-            </Button>
-            <Button h="1.75rem" size="sm" onClick={() => push('/signup')}>
-              Sign Up
-            </Button>
-          </HStack>
-        )}
-      </header>
+      <Nav />
       {allSetup && (
         <Switch>
           <Route exact path="/login" component={Login} />
